requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pypi_0
accelerate=1.2.1=pypi_0
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20250224=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
av=14.0.1=pypi_0
blake3=1.0.4=pypi_0
ca-certificates=2025.8.3=hbd8a1cb_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.3=pyhe01879c_0
compressed-tensors=0.9.1=pypi_0
contourpy=1.3.0=pypi_0
cupy-cuda12x=13.4.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.16=py39haef64b4_0
decorator=5.2.1=pyhd8ed1ab_0
depyf=0.18.0=pypi_0
dill=0.3.9=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
einops=0.8.0=pypi_0
email-validator=2.2.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pyhd8ed1ab_0
fastapi=0.115.10=pypi_0
fastapi-cli=0.0.7=pypi_0
fastrlock=0.8.3=pypi_0
filelock=3.16.1=pypi_0
flash-attn=2.7.4.post1=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
ftfy=6.3.1=pypi_0
gguf=0.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
grpcio=1.75.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
iniconfig=2.0.0=pypi_0
interegular=0.3.3=pypi_0
ipykernel=6.30.1=pyh82676e8_0
ipython=8.18.1=pyh707e725_3
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h723845a_4
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_5
libgcc-ng=15.1.0=h69a702a_5
libgomp=15.1.0=h767d61c_5
libkrb5=1.21.3=h520c7b4_4
libsodium=1.0.20=h4ab18f5_0
libstdcxx=15.1.0=h8f9b012_5
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.43.0=pypi_0
lm-format-enforcer=0.10.11=pypi_0
lmdb=0.9.31=hd590300_1
markdown=3.9=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mistral-common=1.5.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
onnx=1.17.0=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.65.1=pypi_0
openai-clip=1.0.1=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.5.3=h26f9b46_1
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
packaging=24.2=pypi_0
parso=0.8.4=pyhd8ed1ab_1
partial-json-parser=0.2.1.1.post5=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=24.2=py39h06a4308_0
platformdirs=4.3.8=pyhe01879c_0
pluggy=1.5.0=pypi_0
prometheus-client=0.21.1=pypi_0
prometheus-fastapi-instrumentator=7.0.2=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.0=pypi_0
protobuf=5.29.3=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
pybind11=2.13.6=pypi_0
pycocotools=2.0.8=pypi_0
pycountry=24.6.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.1=pypi_0
pytest=8.3.4=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pyhe01879c_2
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.20=pypi_0
python_abi=3.9=2_cp39
pyyaml=6.0.2=pypi_0
pyzmq=26.2.1=pypi_0
qwen-vl-utils=0.0.8=pypi_0
ray=2.40.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rich-toolkit=0.13.2=pypi_0
rpds-py=0.23.1=pypi_0
safetensors=0.5.0=pypi_0
segment-anything=1.0=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.39.0=pypi_0
setuptools=75.1.0=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhe01879c_1
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.46.0=pypi_0
sympy=1.13.1=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tiktoken=0.9.0=pypi_0
timm=1.0.20=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.5.2=py39hd399759_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.49.0=pypi_0
triton=3.1.0=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2024b=h04d1e81_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
vllm=0.7.3=pypi_0
wandb=0.22.0=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websockets=15.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py39h06a4308_0
xformers=0.0.28.post3=pypi_0
xgrammar=0.1.11=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zeromq=4.3.5=h387f397_9
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
